Rebased
Removed shared model.
redefined models in each service.
Removed broker layer, as a quick solution.
Removed redundent admin pages from blazor.
Removed generated Authentication in blazor.
Commented out unused method in model in orders service.

+++ BrokerService/Broker/Warehouse/ProductBroker.cs
@@ -1,7 +1,7 @@
 ﻿using System.Text.Json;
 using ConsoleApp1.Services;
+using sep3.broker.Model;
 using sep3.DTO.Product;
-using sep3.Model;
 
 namespace brokers.broker;
 
@@ -41,7 +41,7 @@ public class ProductBroker : IProductBroker
                 });
                 if (product != null)
                 {
-                    return Result<int>.Success((int)product.Id, "Product created successfully.");
+                    return Result<int>.Success(product.Id, "Product created successfully.");
                 }
                 else
                 {
